# internal/agentd/handlers_specialists.go
package agentd

import (
	"context"
	"encoding/json"
	"maps"
	"net/http"
	"strings"
	"time"

	"github.com/rs/zerolog/log"

	"manifold/internal/config"
	llmproviders "manifold/internal/llm/providers"
	persist "manifold/internal/persistence"
	"manifold/internal/specialists"
	"manifold/internal/tools"
	agenttools "manifold/internal/tools/agents"
	tooldiscovery "manifold/internal/tools/discovery"
)

func (a *app) statusHandler() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		userID, err := a.requireUserID(r)
		if err != nil {
			if a.cfg.Auth.Enabled {
				w.Header().Set("WWW-Authenticate", "Bearer realm=\"sio\"")
			}
			http.Error(w, "unauthorized", http.StatusUnauthorized)
			return
		}
		if r.Method != http.MethodGet {
			http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
			return
		}
		w.Header().Set("Access-Control-Allow-Origin", "*")
		type agentStatus struct {
			ID        string `json:"id"`
			Name      string `json:"name"`
			State     string `json:"state"`
			Model     string `json:"model"`
			UpdatedAt string `json:"updatedAt"`
		}
		list, err := a.specStore.List(r.Context(), userID)
		if err != nil {
			http.Error(w, "internal server error", http.StatusInternalServerError)
			return
		}
		now := time.Now().UTC().Format(time.RFC3339)
		out := make([]agentStatus, 0, len(list))
		for _, s := range list {
			if s.Paused {
				// Maintain previous behaviour: paused specialists are hidden
				// from the Overview cards so they read as "offline".
				continue
			}
			out = append(out, agentStatus{
				ID:        s.Name,
				Name:      s.Name,
				State:     "online",
				Model:     s.Model,
				UpdatedAt: now,
			})
		}
		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(out)
	}
}

func (a *app) specialistDefaultsHandler() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if _, err := a.requireUserID(r); err != nil {
			if a.cfg.Auth.Enabled {
				w.Header().Set("WWW-Authenticate", "Bearer realm=\"sio\"")
			}
			http.Error(w, "unauthorized", http.StatusUnauthorized)
			return
		}
		if r.Method != http.MethodGet {
			http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		out := map[string]persist.Specialist{}
		for _, p := range []string{"openai", "anthropic", "google", "local"} {
			model, baseURL, apiKey, headers, params := a.providerDefaults(p)
			out[p] = persist.Specialist{
				Provider:     p,
				BaseURL:      baseURL,
				APIKey:       apiKey,
				Model:        model,
				ExtraHeaders: headers,
				ExtraParams:  params,
			}
		}
		json.NewEncoder(w).Encode(out)
	}
}

func (a *app) specialistsHandler() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		userID, err := a.requireUserID(r)
		if err != nil {
			if a.cfg.Auth.Enabled {
				w.Header().Set("WWW-Authenticate", "Bearer realm=\"sio\"")
			}
			http.Error(w, "unauthorized", http.StatusUnauthorized)
			return
		}
		switch r.Method {
		case http.MethodGet:
			list, err := a.listSpecialistsForUser(r.Context(), userID)
			if err != nil {
				http.Error(w, "internal server error", http.StatusInternalServerError)
				return
			}
			w.Header().Set("Content-Type", "application/json")
			json.NewEncoder(w).Encode(list)

		case http.MethodPost:
			r.Body = http.MaxBytesReader(w, r.Body, 1<<20)
			defer r.Body.Close()
			var sp persist.Specialist
			if err := json.NewDecoder(r.Body).Decode(&sp); err != nil {
				http.Error(w, "bad request", http.StatusBadRequest)
				return
			}
			saved, status, err := a.createSpecialistForUser(r.Context(), userID, sp)
			if err != nil {
				http.Error(w, err.Error(), http.StatusBadRequest)
				return
			}
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(status)
			json.NewEncoder(w).Encode(saved)

		default:
			http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
		}
	}
}

func (a *app) specialistDetailHandler() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		userID, err := a.requireUserID(r)
		if err != nil {
			if a.cfg.Auth.Enabled {
				w.Header().Set("WWW-Authenticate", "Bearer realm=\"sio\"")
			}
			http.Error(w, "unauthorized", http.StatusUnauthorized)
			return
		}
		name := strings.TrimPrefix(r.URL.Path, "/api/specialists/")
		name = strings.TrimSpace(name)
		if name == "" {
			http.NotFound(w, r)
			return
		}

		switch r.Method {
		case http.MethodGet:
			sp, ok, err := a.getSpecialistForUser(r.Context(), userID, name)
			if err != nil {
				http.Error(w, "internal server error", http.StatusInternalServerError)
				return
			}
			if !ok {
				http.NotFound(w, r)
				return
			}
			w.Header().Set("Content-Type", "application/json")
			json.NewEncoder(w).Encode(sp)
		case http.MethodPut:
			r.Body = http.MaxBytesReader(w, r.Body, 1<<20)
			defer r.Body.Close()
			var sp persist.Specialist
			if err := json.NewDecoder(r.Body).Decode(&sp); err != nil {
				http.Error(w, "bad request", http.StatusBadRequest)
				return
			}
			saved, err := a.updateSpecialistForUser(r.Context(), userID, name, sp)
			if err != nil {
				http.Error(w, err.Error(), http.StatusBadRequest)
				return
			}
			w.Header().Set("Content-Type", "application/json")
			json.NewEncoder(w).Encode(saved)
		case http.MethodDelete:
			if err := a.deleteSpecialistForUser(r.Context(), userID, name); err != nil {
				if err == errOrchestratorDelete {
					http.Error(w, err.Error(), http.StatusBadRequest)
					return
				}
				http.Error(w, "error", http.StatusInternalServerError)
				return
			}
			w.WriteHeader(http.StatusNoContent)
		default:
			http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
		}
	}
}

func (a *app) orchestratorSpecialist(ctx context.Context, userID int64) persist.Specialist {
	defaultProvider := strings.TrimSpace(a.cfg.LLMClient.Provider)
	if defaultProvider == "" {
		defaultProvider = "openai"
	}
	baseModel, baseURL, baseKey, baseHeaders, baseParams := a.providerDefaults(defaultProvider)
	out := persist.Specialist{
		ID:                         0,
		UserID:                     userID,
		Name:                       specialists.OrchestratorName,
		Description:                "",
		Provider:                   defaultProvider,
		BaseURL:                    baseURL,
		APIKey:                     baseKey,
		Model:                      baseModel,
		SummaryContextWindowTokens: 0,
		EnableTools:                a.cfg.EnableTools,
		RequestInfoEnabled:         boolPtr(config.RequestInfoEnabled(a.cfg.RequestInfoEnabled)),
		AutoDiscover:               boolPtr(a.cfg.AutoDiscover),
		Paused:                     false,
		AllowTools:                 a.cfg.ToolAllowList,
		System:                     a.orchestratorSystemPrompt(),
		ExtraHeaders:               baseHeaders,
		ExtraParams:                baseParams,
	}
	// Apply per-user overlay if present
	if sp, ok, _ := a.specStore.GetByName(ctx, userID, specialists.OrchestratorName); ok {
		out.ID = sp.ID
		out.Description = sp.Description
		if strings.TrimSpace(sp.Provider) != "" {
			out.Provider = strings.TrimSpace(sp.Provider)
			// Refresh defaults for the saved provider
			out.Model, out.BaseURL, out.APIKey, out.ExtraHeaders, out.ExtraParams = a.providerDefaults(out.Provider)
		}
		if strings.TrimSpace(sp.BaseURL) != "" {
			out.BaseURL = sp.BaseURL
		}
		if strings.TrimSpace(sp.APIKey) != "" {
			out.APIKey = sp.APIKey
		}
		if strings.TrimSpace(sp.Model) != "" {
			out.Model = sp.Model
		}
		if sp.SummaryContextWindowTokens > 0 {
			out.SummaryContextWindowTokens = sp.SummaryContextWindowTokens
		}
		out.EnableTools = sp.EnableTools
		if sp.RequestInfoEnabled != nil {
			out.RequestInfoEnabled = boolPtr(*sp.RequestInfoEnabled)
		}
		if sp.AutoDiscover != nil {
			out.AutoDiscover = boolPtr(*sp.AutoDiscover)
		}
		if sp.AllowTools != nil {
			out.AllowTools = append([]string(nil), sp.AllowTools...)
		}
		out.ReasoningEffort = sp.ReasoningEffort
		if strings.TrimSpace(sp.System) != "" {
			out.System = sp.System
		}
		if sp.ExtraHeaders != nil {
			out.ExtraHeaders = sp.ExtraHeaders
		}
		if sp.ExtraParams != nil {
			out.ExtraParams = copyAnyMap(sp.ExtraParams)
		}
		if sp.Harness != nil {
			out.Harness = sp.Harness
		}
	}
	return out
}

func (a *app) providerDefaults(provider string) (model, baseURL, apiKey string, headers map[string]string, params map[string]any) {
	switch provider {
	case "anthropic":
		baseURL = strings.TrimSpace(a.cfg.LLMClient.Anthropic.BaseURL)
		apiKey = strings.TrimSpace(a.cfg.LLMClient.Anthropic.APIKey)
		model = strings.TrimSpace(a.cfg.LLMClient.Anthropic.Model)
		params = copyAnyMap(a.cfg.LLMClient.Anthropic.ExtraParams)
	case "google":
		baseURL = strings.TrimSpace(a.cfg.LLMClient.Google.BaseURL)
		apiKey = strings.TrimSpace(a.cfg.LLMClient.Google.APIKey)
		model = strings.TrimSpace(a.cfg.LLMClient.Google.Model)
		params = copyAnyMap(a.cfg.LLMClient.Google.ExtraParams)
	default:
		baseURL = strings.TrimSpace(a.cfg.LLMClient.OpenAI.BaseURL)
		apiKey = strings.TrimSpace(a.cfg.LLMClient.OpenAI.APIKey)
		model = strings.TrimSpace(a.cfg.LLMClient.OpenAI.Model)
		headers = copyStringMap(a.cfg.LLMClient.OpenAI.ExtraHeaders)
		params = copyAnyMap(a.cfg.LLMClient.OpenAI.ExtraParams)
	}
	if headers == nil {
		headers = map[string]string{}
	}
	if params == nil {
		params = map[string]any{}
	}
	return model, baseURL, apiKey, headers, params
}

func (a *app) teamMembershipsForUser(ctx context.Context, userID int64) map[string][]string {
	if a.teamStore == nil {
		return map[string][]string{}
	}
	m, err := a.teamStore.ListMemberships(ctx, userID)
	if err != nil {
		return map[string][]string{}
	}
	return m
}

func (a *app) applyTeamMemberships(ctx context.Context, userID int64, specialistName string, teams []string) error {
	if a.teamStore == nil {
		return nil
	}
	if teams == nil {
		return nil
	}
	current := a.teamMembershipsForUser(ctx, userID)[specialistName]
	currentSet := map[string]struct{}{}
	for _, t := range current {
		currentSet[t] = struct{}{}
	}
	desiredSet := map[string]struct{}{}
	for _, t := range teams {
		t = strings.TrimSpace(t)
		if t == "" {
			continue
		}
		desiredSet[t] = struct{}{}
	}
	for t := range currentSet {
		if _, ok := desiredSet[t]; !ok {
			if err := a.teamStore.RemoveMember(ctx, userID, t, specialistName); err != nil {
				return err
			}
		}
	}
	for t := range desiredSet {
		if _, ok := currentSet[t]; !ok {
			if err := a.teamStore.AddMember(ctx, userID, t, specialistName); err != nil {
				return err
			}
		}
	}
	return nil
}

func (a *app) removeSpecialistFromTeams(ctx context.Context, userID int64, specialistName string) error {
	if a.teamStore == nil {
		return nil
	}
	current := a.teamMembershipsForUser(ctx, userID)[specialistName]
	for _, t := range current {
		if err := a.teamStore.RemoveMember(ctx, userID, t, specialistName); err != nil {
			return err
		}
	}
	return nil
}

func copyStringMap(in map[string]string) map[string]string {
	if len(in) == 0 {
		return map[string]string{}
	}
	out := make(map[string]string, len(in))
	maps.Copy(out, in)
	return out
}

func copyAnyMap(in map[string]any) map[string]any {
	if len(in) == 0 {
		return map[string]any{}
	}
	out := make(map[string]any, len(in))
	maps.Copy(out, in)
	return out
}

func (a *app) applyOrchestratorUpdate(ctx context.Context, sp persist.Specialist) error {
	provider := specialists.ApplyOrchestratorConfig(a.cfg, sp)

	llm, err := llmproviders.Build(*a.cfg, a.httpClient)
	if err != nil {
		return err
	}
	a.llm = llm
	a.engine.LLM = llm
	a.engine.Model = a.orchestratorModel(provider, sp.Model)
	a.engine.HarnessEnabled = a.cfg.Harness.Enabled
	a.engine.HarnessConfig = harnessRunConfig(a.cfg.Harness)

	a.refreshOrchestratorToolRegistry()
	a.engine.Tools = a.toolRegistry
	// Propagate updated tool registry to the delegator (if present)
	if a.engine != nil && a.engine.Delegator != nil {
		if d, ok := a.engine.Delegator.(*agenttools.Delegator); ok {
			d.SetRegistry(a.toolRegistry)
		}
	}

	toSave := persist.Specialist{
		Name:                       specialists.OrchestratorName,
		Description:                sp.Description,
		EnableTools:                a.cfg.EnableTools,
		RequestInfoEnabled:         boolPtr(config.RequestInfoEnabled(a.cfg.RequestInfoEnabled)),
		AutoDiscover:               boolPtr(a.cfg.AutoDiscover),
		Paused:                     false,
		AllowTools:                 append([]string(nil), a.cfg.ToolAllowList...),
		System:                     a.orchestratorSystemPrompt(),
		Provider:                   provider,
		ExtraParams:                sp.ExtraParams,
		SummaryContextWindowTokens: sp.SummaryContextWindowTokens,
		Harness:                    sp.Harness,
	}
	switch provider {
	case "anthropic":
		toSave.BaseURL = a.cfg.LLMClient.Anthropic.BaseURL
		toSave.APIKey = a.cfg.LLMClient.Anthropic.APIKey
		toSave.Model = a.cfg.LLMClient.Anthropic.Model
	case "google":
		toSave.BaseURL = a.cfg.LLMClient.Google.BaseURL
		toSave.APIKey = a.cfg.LLMClient.Google.APIKey
		toSave.Model = a.cfg.LLMClient.Google.Model
	default:
		toSave.BaseURL = a.cfg.LLMClient.OpenAI.BaseURL
		toSave.APIKey = a.cfg.LLMClient.OpenAI.APIKey
		toSave.Model = a.cfg.LLMClient.OpenAI.Model
		toSave.ExtraHeaders = a.cfg.LLMClient.OpenAI.ExtraHeaders
		toSave.ReasoningEffort = sp.ReasoningEffort
	}
	toSave.UserID = systemUserID
	if _, err := a.specStore.Upsert(ctx, systemUserID, toSave); err != nil {
		log.Error().Err(err).Msg("failed to persist orchestrator configuration")
		return err
	}
	if list, err := a.specStore.List(ctx, systemUserID); err == nil {
		a.specRegistry.ReplaceFromConfigs(a.cfg.LLMClient, specialists.ConfigsFromStore(list), a.httpClient, a.baseToolRegistry)
		a.specRegistry.SetPromptOverrides(promptInstructionOverrides(a.cfg))
		a.specRegistry.SetRequestInfoEnabled(config.RequestInfoEnabled(a.cfg.RequestInfoEnabled))
		a.specRegistry.SetToolDiscovery(a.toolIndex, a.cfg.AutoDiscover, a.cfg.MaxDiscoveredTools)
	}
	a.refreshEngineSystemPrompt()
	names := make([]string, 0, len(a.toolRegistry.Schemas()))
	for _, s := range a.toolRegistry.Schemas() {
		names = append(names, s.Name)
	}
	log.Info().
		Bool("enableTools", a.cfg.EnableTools).
		Strs("allowList", a.cfg.ToolAllowList).
		Strs("tools", names).
		Msg("tool_registry_contents_updated")
	return nil
}

func (a *app) orchestratorModel(provider string, override string) string {
	if currentModel := strings.TrimSpace(override); currentModel != "" {
		return currentModel
	}
	switch provider {
	case "anthropic":
		return strings.TrimSpace(a.cfg.LLMClient.Anthropic.Model)
	case "google":
		return strings.TrimSpace(a.cfg.LLMClient.Google.Model)
	default:
		return strings.TrimSpace(a.cfg.LLMClient.OpenAI.Model)
	}
}

func (a *app) refreshOrchestratorToolRegistry() {
	if a.cfg.AutoDiscover && a.cfg.EnableTools && a.toolIndex != nil {
		a.toolRegistry = tooldiscovery.NewDiscoverableRegistry(a.baseToolRegistry, a.toolIndex, a.cfg.ToolAllowList, a.cfg.MaxDiscoveredTools)
	} else {
		a.toolRegistry = tools.ApplyTopLevelPolicy(a.baseToolRegistry, a.cfg.EnableTools, a.cfg.ToolAllowList)
	}
	a.toolRegistry = withChatInputRequestTool(a.toolRegistry, a.cfg.EnableTools && config.RequestInfoEnabled(a.cfg.RequestInfoEnabled))
}